создает отдельный модуль "renderComments"
